feat: 검증 로직을 별도 파일로 분리
