+++ src/calculator.js
@@ -1,14 +1,11 @@
 import {Console} from '@woowacourse/mission-utils';
+import {validateAndConvertNumber, validateCustomSeparator, validateInput} from "./validators.js";
 
 export async function calculator() {
-    try {
         const input = await Console.readLineAsync('덧셈할 문자열을 입력해 주세요.')
         const parseResult = parseInput(input);
         const result = add(parseResult);
         Console.print(`결과 : ${result}`);
-    } catch (error) {
-        throw error;
-    }
 }
 
 function parseInput(input) {
@@ -35,26 +32,4 @@ function add(parseResult) {
     return parseResult.reduce((acc, cur) => acc + cur, 0)
 }
 
-function validateInput(input) {
-    if (input.trim() === '') {
-        throw Error("[ERROR]");
-    }
-}
-
-function validateCustomSeparator(input,number ) {
-    if(!input.startsWith('//') || !number) {
-        throw Error("[ERROR]")
-    }
-}
-
-function validateAndConvertNumber(item) {
-    const num = Number(item);
-    if (isNaN(num)) {
-        throw Error("[ERROR]");
-    }
-    if (num < 0) {
-        throw Error("[ERROR]");
-    }
-    return num;
-}
